pyproject.toml
[tool.poetry]
name = "pycsvy"
version = "0.1.0"
description = "Library that implement saving/loading together tabular data and yaml-formatted metadata."
authors = ["Diego <d.alonso-alvarez@imperial.ac.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-cov = "^3.0.0"
pytest-flake8 = "^1.1.1"
pytest-mypy = "^0.9.1"
pytest-mock = "^3.7.0"
isort = "^5.10.1"
pre-commit = "^2.18.1"
black = "^22.3.0"
numpy = "^1.22.3"
pandas = "^1.4.1"
flake8 = "^4.0.1"
mypy = "^0.950"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
